Port the "to satisfy" test suite.
Swap all cases that tested object comparisons with Map comparisons to
ensure the Map centric copy of "to satisfy" is functionally in line with
the original for object.

The port was done by copying the file and changing all tests on objects to
tests on Map. While some blocks were removed the order of the file was not
changed - this was attempt to ensure some chance of 'em being kept in sync.