Rework "to satisfy" diffs so they are always executable.
The Unexpected convention is that output should be directly copyable
into source code or the repl. The plugin did not previously adhere
to that standard, primarily because entires in a Map have a closing
square bracket but object version, from which this was ported, would
just write messages at the end of the line with // comments. Address
this by properly accommodating that we need additional newlines.